bugs and cleanup

+++ helpers/helpers.sh
@@ -100,18 +100,20 @@ cleaner() {
   # now we can unmount the stuff from emulator and delete temporary stuff
 
   # if S115 is found only once in main.log the module was started and we have to clean it up
-  if [[ $(grep -c S115 "$LOG_DIR"/"$MAIN_LOG_FILE") -eq 1 ]]; then
-    print_output "[*] Terminating qemu processes - check it with ps" "no_log"
-    killall -9 --quiet -r .*qemu.*sta.*
-    print_output "[*] Cleaning the emulation environment\\n" "no_log"
-    find "$FIRMWARE_PATH_CP" -xdev -iname "qemu*static" -exec rm {} \; 2>/dev/null
-    print_output "[*] Umounting proc, sys and run" "no_log"
-    mapfile -t CHECK_MOUNTS < <(mount | grep "$FIRMWARE_PATH_CP")
-    for MOUNT in "${CHECK_MOUNTS[@]}"; do
-      print_output "[*] Unmounting $MOUNT" "no_log"
-      MOUNT=$(echo "$MOUNT" | cut -d\  -f3)
-      umount -l "$MOUNT"
-    done
+  if [[ -f "$LOG_DIR"/"$MAIN_LOG_FILE" ]]; then
+    if [[ $(grep -c S115 "$LOG_DIR"/"$MAIN_LOG_FILE") -eq 1 ]]; then
+      print_output "[*] Terminating qemu processes - check it with ps" "no_log"
+      killall -9 --quiet -r .*qemu.*sta.*
+      print_output "[*] Cleaning the emulation environment\\n" "no_log"
+      find "$FIRMWARE_PATH_CP" -xdev -iname "qemu*static" -exec rm {} \; 2>/dev/null
+      print_output "[*] Umounting proc, sys and run" "no_log"
+      mapfile -t CHECK_MOUNTS < <(mount | grep "$FIRMWARE_PATH_CP")
+      for MOUNT in "${CHECK_MOUNTS[@]}"; do
+        print_output "[*] Unmounting $MOUNT" "no_log"
+        MOUNT=$(echo "$MOUNT" | cut -d\  -f3)
+        umount -l "$MOUNT"
+      done
+    fi
   fi
 
   if [[ -d "$TMP_DIR" ]]; then
